Split ace_tracker.py into ace_data.py / ace_html.py / ace_tracker.py
Pure refactor, no behavior change — ace_tracker.py had grown to ~2,600
lines mixing data fetch/calculation, HTML rendering, and orchestration.
Splitting now makes it easier to place the generation code that #51
(individual storm pages) and #63 (all-time records page) will add.

Every moved function/constant was verified byte-identical to its
pre-refactor source, and an end-to-end run before/after the split
produced HTML output that differs only in the "Updated" timestamp.

Closes #52
